+++ lab_2/.idea/workspace.xml
@@ -12,9 +12,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="2296888d-df11-4f1b-8a67-d52e5ef1fdb5" name="Default" comment="">
+      <change type="MOVED" beforePath="$PROJECT_DIR$/../lab_3/main_dilu.c" afterPath="$PROJECT_DIR$/../lab_3/main.c" />
       <change type="MODIFICATION" beforePath="$PROJECT_DIR$/.idea/workspace.xml" afterPath="$PROJECT_DIR$/.idea/workspace.xml" />
       <change type="MODIFICATION" beforePath="$PROJECT_DIR$/../lab_3/.idea/workspace.xml" afterPath="$PROJECT_DIR$/../lab_3/.idea/workspace.xml" />
-      <change type="MODIFICATION" beforePath="$PROJECT_DIR$/../lab_3/main_dilu.c" afterPath="$PROJECT_DIR$/../lab_3/main_dilu.c" />
+      <change type="MODIFICATION" beforePath="$PROJECT_DIR$/../lab_3/Makefile" afterPath="$PROJECT_DIR$/../lab_3/Makefile" />
+      <change type="MODIFICATION" beforePath="$PROJECT_DIR$/../lab_3/data/main.c" afterPath="$PROJECT_DIR$/../lab_3/data/main.c" />
     </list>
     <ignored path="proj1.iws" />
     <ignored path=".idea/workspace.xml" />
