+++ lab_3/data/main.c
@@ -28,14 +28,14 @@ int main(){
     int NY = 180;
     int NZ = 38;
     int size = NX * NY * NZ;
-    /*
+
     char file_name[100] = "./case_360x180x38/A.txt";
     FILE *fp=fopen(file_name,"r");
     if(fp==NULL){
         printf("can not open file!\n");
         exit(0);
     }
-    around_point *data = (around_point *)malloc(size * sizeof(around_point));
+    double *data = (double *)malloc(size * 19 * sizeof(double));
     int i,j,k,num;
     double val;
     int index = 0;
@@ -47,11 +47,8 @@ int main(){
             fscanf(fp,"%d",&i); fscanf(fp,"%d",&j);
             fscanf(fp,"%d",&k); fscanf(fp,"%d",&num);
             fscanf(fp,"%lf",&val);
-            data[index].val[sub_index] = val;
+            data[19 * index + sub_index] = val;
             if (sub_index == 0){
-                data[index].i = i;
-                data[index].j = j;
-                data[index].k = k;
                 if (count == 10000){
                     printf("i: %d, j : %d, k : %d \n", i, j, k);
                     count = 0;
@@ -62,22 +59,23 @@ int main(){
         count++;
     }
     fclose(fp);
-    FILE *fw=fopen("data_A.bin","wb");
+    FILE *fw=fopen("data_A_v1.bin","wb");
     if(fw==NULL){
         printf("can't open the write file\n");
         exit(0);
     }
-    fwrite(data, sizeof(around_point), size, fw);
+    fwrite(data, sizeof(double), size * 19, fw);
     fclose(fw);
     free(data);
-    */
-    char file_name[100] = "./case_360x180x38/x0.txt";
+
+    /*
+    char file_name[100] = "./case_360x180x38/b.txt";
     FILE *fp=fopen(file_name,"r");
     if(fp==NULL){
         printf("can not open file!\n");
         exit(0);
     }
-    b_struct *data = (b_struct *)malloc(size * sizeof(b_struct));
+    double *data = (double *)malloc(size * sizeof(double));
     int i,j,k;
     double val;
     int index = 0;
@@ -88,10 +86,7 @@ int main(){
         fscanf(fp,"%d",&j);
         fscanf(fp,"%d",&k);
         fscanf(fp,"%lf",&val);
-        data[index].i = i;
-        data[index].j = j;
-        data[index].k = k;
-        data[index].val = val;
+        data[index] = val;
         count++;
         if (count == 10000){
             printf("i : %d, j : %d, k : %d, val : %.12lf \n", i, j, k, val);
@@ -99,13 +94,14 @@ int main(){
         }
     }
     fclose(fp);
-    FILE *fw=fopen("./case_1bin/data_x0.bin","wb");
+    FILE *fw=fopen("./case_1bin/data_b_v1.bin","wb");
     if(fw==NULL){
         printf("can't open the write file\n");
         exit(0);
     }
-    fwrite(data, sizeof(b_struct), size, fw);
+    fwrite(data, sizeof(double), size, fw);
     fclose(fw);
     free(data);
+     */
     return 0;
 }
