+++ lab_3/main.c
@@ -4,6 +4,7 @@
 #include <unistd.h>
 #include <sys/fcntl.h>
 #include <string.h>
+#include <mpi.h>
 
 #define TIME(a, b) (1.0*((b).tv_sec-(a).tv_sec)+0.000001*((b).tv_usec-(a).tv_usec))
 typedef long long ll;
@@ -28,12 +29,6 @@ typedef struct {
   int len_x;
 } Data_Info;
 
-typedef struct{
-  int i;
-  int j;
-  int k;
-  double val[19];
-} around_point;
 
 typedef struct{
   int col;
@@ -44,8 +39,16 @@ typedef struct{
   int i;
   int j;
   int k;
+  double val[19];
+} around_point;
+
+typedef struct{
+  int i;
+  int j;
+  int k;
   double val;
 } v_struct;
+
 
 double *A_value; // partial A
 int *A_col;
@@ -53,37 +56,51 @@ int *A_ptr;
 double *b; // partial b
 double *x; // partial x, recieve buffer is behind the local x
 double *send_buffer;//store the data to send, 8 segments, not all are used
-//MPI_Request recv_request[8]; // receive request to check needed data has been received. not all are used
-//MPI_Request send_request[8]; // send requets
+int *g_map;//map behind index to global index
+MPI_Request recv_request[8]; // receive request to check needed data has been received. not all are used
+MPI_Request send_request[8]; // send requets
 int buffer_offset[8]; // the start index of par i to send
 int total_nodes;
 int node_dsize[8]; // this node communicates with 8 nodes at most, this array save the date size to transfer
-int node_rank[8]; // save the node rank to communicate
-int node_offset[8];// record the offset data to receive and send
+int node_rank[8]; // save the node rank to communication
+int target_side[8]; // 初始化目标的index, 从接受者的角度来看
+// receive buffer is stored in R_hat
 
 int x_cor[19] = {0, -1, +1, 0, 0, +1, +1, -1, -1, 0, -1, +1, 0, 0, 0, -1, +1, 0, 0};
 int y_cor[19] = {0, 0, 0, -1, +1, +1, -1, -1, +1, 0, 0, 0, -1, +1, 0, 0, 0, -1, +1};
 int z_cor[19] = {0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -1, -1, -1, -1, +1, +1, +1, +1, +1};
 
+int min(int a, int b){
+    if (a > b){
+        return b;
+    } else {
+        return a;
+    }
+}
+
 Data_Info init_info(int NX, int NY, int NZ, int PX, int PY, int PZ){
     Data_Info info;
-    if (PX == 1, PY == 1, PZ == 1){
+    info.tX = NX;
+    info.tY = NY;
+    info.tZ = NZ;
+    info.x_par = PX;
+    info.y_par = PY;
+    info.z_par = PZ;
+    info.nz = NZ;
+    info.z_start = 0;
+    info.z_end = NZ;
+    int myrank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &myrank);
+    info.myrank = myrank;
+    int sub_i;
+    if (PX == 1 && PY == 1 && PZ == 1){
         //info.myrank
         info.nx = NX;
         info.ny = NY;
-        info.nz = NZ;
-        info.x_par = PX;
-        info.y_par = PY;
-        info.z_par = PZ;
         info.x_start = 0;
         info.x_end = NX;
         info.y_start = 0;
         info.y_end = NY;
-        info.z_start = 0;
-        info.z_end = NZ;
-        info.tX = NX;
-        info.tY = NY;
-        info.tZ = NZ;
         ll size_col = (ll) NX * (ll) NY * (ll) NZ;
         ll size_a = size_col * 19;
         A_value = (double *) malloc(size_a * sizeof(double));
@@ -95,6 +112,10 @@ Data_Info init_info(int NX, int NY, int NZ, int PX, int PY, int PZ){
         int size_buffer = 4 * NZ + 2 * NZ * NY + 2 * NZ * NX;
         //receive_buffer = (double *) malloc(size_buffer * sizeof(double));
         info.len_x = info.len_vb + size_buffer;
+        g_map = (int *) malloc(size_buffer * sizeof(int));
+        for (sub_i = 0; sub_i < size_buffer; sub_i++){
+            g_map[sub_i] = -1;
+        }
         x = (double *) malloc(info.len_x * sizeof(double));
         memset(x, 0, info.len_x * sizeof(double));
         send_buffer = (double *) malloc(size_buffer * sizeof(double));
@@ -102,13 +123,129 @@ Data_Info init_info(int NX, int NY, int NZ, int PX, int PY, int PZ){
         int i;
         for (i = 0; i < 8; i++){
             node_dsize[i] = -1;
-            node_rank[i] = -1;
+            node_rank[i] = 0;
+            buffer_offset[i] = -1;
         }
+        return info;
+    } else {
+        // 首先切 Y 轴, naive
+        int par_y_index = myrank / PX;
+        int interval_y = NY / PY + 1;
+        info.y_start = par_y_index * interval_y;
+        info.y_end = min(NY, info.y_start + interval_y);
+        info.ny = info.y_end - info.y_start;
+        // 切 X 轴, 比较繁琐
+        int par_x_index = myrank % PX;
+        int bi_par_x = PX / 2;
+        int bi_nx = NX / 2;
+        int interval_x = bi_nx / bi_par_x;
+        if (par_x_index < bi_par_x){
+            info.x_start = interval_x * par_x_index;
+            info.x_end = min(bi_nx, info.x_start + interval_x);
+        } else {
+            info.x_start = bi_nx + interval_x * (par_x_index - bi_par_x);
+            info.x_end = min(NX, info.x_start + interval_x);
+            //printf("x_start: %d, x_end : %d \n", info.x_start, info.x_end);
+        }
+        info.nx = info.x_end - info.x_start;
+    }
+    int size_col = info.nx * info.ny * info.nz;
+    int size_a = size_col * 19;
+    A_value = (double *) malloc(size_a * sizeof(double));
+    A_col = (int *) malloc(size_a * sizeof(int));
+    A_ptr = (int *) malloc((size_col + 1) * sizeof(int));
+    //printf("nx : %d, ny : %d, nz : %d\n", info.nx, info.ny, info.nz);
+    b = (double *) malloc(size_col * sizeof(double));
+    info.len_vb = (int)size_col;
+    // cal buffer size
+    int size_buffer = 4 * info.nz + 2 * info.nz * info.ny + 2 * info.nz * info.nx;
+    g_map = (int *) malloc(size_buffer * sizeof(int));
+    for (sub_i = 0; sub_i < size_buffer; sub_i++){
+        g_map[sub_i] = -1;
+    }
+    if (g_map == NULL){
+        printf("null\n");
+    }
+    //receive_buffer = (double *) malloc(size_buffer * sizeof(double));
+    info.len_x = info.len_vb + size_buffer;
+    x = (double *) malloc(info.len_x * sizeof(double));
+    for (sub_i = 0; sub_i < info.len_x; sub_i++){
+        x[sub_i] = -2;
+    }
+    memset(x, 0, info.len_x * sizeof(double));
+    send_buffer = (double *) malloc(size_buffer * sizeof(double));
+    //init buffer start index and buffer size, same to send and receive
+    buffer_offset[0] = 0;
+    buffer_offset[1] = info.nz + buffer_offset[0];
+    buffer_offset[2] = info.nz * info.ny + buffer_offset[1];
+    buffer_offset[3] = info.nz + buffer_offset[2];
+    buffer_offset[4] = info.nz * info.nx + buffer_offset[3];
+    buffer_offset[5] = info.nz + buffer_offset[4];
+    buffer_offset[6] = info.nz * info.ny + buffer_offset[5];
+    buffer_offset[7] = info.nz + buffer_offset[6];
+    int i;
+    for(i = 0; i < 7; i++){
+        node_dsize[i] = buffer_offset[i + 1] - buffer_offset[i];
+        //printf("node data size[%d] : %d \n", i, node_dsize[i]);
+    }
+    node_dsize[7] = info.nz * info.nx;
+    // init node to send and receive, and the target side from the receiver
+    int line_rank = myrank % PX;
+    int line_lay = myrank / PX;
+    int bi_px = PX / 2;
+    if (info.y_start == 0){
+        //side 0
+        node_rank[0] = (line_rank - 1 + bi_px) % PX + line_lay * PX;
+        target_side[0] = 6; //send
+        // side 7
+        node_rank[7] = (line_rank + bi_px) % PX + line_lay * PX;
+        target_side[7] = 7;
+        //side 6
+        node_rank[6] = (line_rank + 1 + bi_px) % PX + line_lay * PX;
+        target_side[6] = 0;
+    } else {
+        // side 0
+        node_rank[0] = (line_rank - 1 + PX) % PX + line_lay * PX - PX;
+        target_side[0] = 4;
+        // side 7
+        node_rank[7] = myrank - PX;
+        target_side[7] = 3;
+        //side 6
+        node_rank[6] = (line_rank + 1) % PX + line_lay * PX - PX;
+        target_side[6] = 2;
+    }
+    // side 1
+    node_rank[1] = (line_rank - 1 + PX) % PX + line_lay * PX;
+    target_side[1] = 5;
+    // side 5
+    node_rank[5] = (line_rank + 1) % PX + line_lay * PX;
+    target_side[5] = 1;
+    if (info.y_end == NY){
+        //side 2
+        node_rank[2] = (line_rank - 1 + bi_px) % PX + line_lay * PX;
+        target_side[2] = 4; //send
+        // side 3
+        node_rank[3] = (line_rank + bi_px ) % PX + line_lay * PX;
+        target_side[3] = 3;
+        // side 4
+        node_rank[4] = (line_rank + 1 + bi_px) % PX + line_lay * PX;
+        target_side[4] = 2;
+    } else {
+        //side 2
+        node_rank[2] = (line_rank - 1 + PX) % PX + line_lay * PX + PX;
+        target_side[2] = 6;
+        //side 3
+        node_rank[3] = myrank + PX;
+        target_side[3] = 7;
+        //side 4
+        node_rank[4] = (line_rank + 1) % PX + line_lay * PX + PX;
+        target_side[4] = 0;
     }
     return info;
 }
 
-int get_col(int i, int j, int k, Data_Info info){
+int get_col(int i, int j, int k, Data_Info info, int *g_index){
+    *g_index = 0;
     if (i >= 0 && i < info.nx && j >= 0 && j < info.ny && k >=0 && k < info.nz){
         return (i * info.ny * info.nz + j * info.nz + k);
     }
@@ -124,11 +261,11 @@ int get_col(int i, int j, int k, Data_Info info){
     } else {
         if (g_j == info.tY){
             g_j = info.tY - 1;
-            g_i += (info.tY >> 1);
+            g_i += (info.tX >> 1);
         }
     }
     // x 循环回来
-    g_i = (g_i + info.nx) % info.nx;
+    g_i = (g_i + info.tX) % info.tX;
     if (g_i >= info.x_start && g_i < info.x_end && g_j >= info.y_start && g_j < info.y_end && g_k >= info.z_start && g_k < info.z_end){
         // the node is still in cube
         int l_i = g_i - info.x_start;
@@ -138,34 +275,63 @@ int get_col(int i, int j, int k, Data_Info info){
     } else {
         // the node is outof cube
         int size = info.nx * info.ny * info.nz;
-        if (i == -1){
-            return (size + (j + 1) * info.nz + k);
-        } else {
-            if (i == info.nx){
-                return (size + (info.ny + 2) * info.nz + info.nx * info.nz * 2 + (j + 1) * info.nz + k);
+        *g_index = g_i * info.tY * info.tZ + g_j * info.tZ + g_k;
+        if (j == -1){
+            if (i == -1){
+                //side 0
+                return (size + k);
             } else {
-                if (j == -1){
-                    return (size + (info.ny + 2) * info.nz + info.nz * i + k);
+                if (i == info.nx){
+                    //side 6
+                    return (size + info.nz * 3 + info.nz * info.ny * 2 + info.nz * info.nx + k);
                 } else {
-                    return (size + (info.ny + 2) * info.nz + info.nz * info.nx+ info.nz * i + k);
+                    //side 7
+                    return (size + info.nz * 4 + info.nz * info.ny * 2 + info.nz * info.nx + i * info.nz + k);
+                }
+            }
+        } else {
+            if (j == info.ny){
+                if (i == -1){
+                    //side 2
+                    return (size + info.nz + info.nz * info.ny + k);
+                } else {
+                    if (i == info.nx){
+                        //side 4
+                        return (size + info.nz * 2 + info.nz * info.ny + info.nz * info.nx + k);
+                    } else {
+                        //side 3
+                        return (size + info.nz * 2 + info.nz * info.ny + i * info.nz + k);
+                    }
+                }
+            } else {
+                if (i == -1){
+                    //side 1
+                    return (size + info.nz + j * info.nz + k);
+                } else {
+                    //side 5
+                    if (i == info.nx){
+                        return (size + info.nz * 3 + info.nz * info.ny + info.nz * info.nx + j * info.nz + k);
+                    } else {
+                        //printf("error!!!!!!!!!!!!!!!!!!!!\n");
+                        return -1;
+                    }
+
                 }
             }
         }
     }
 }
 
-int pair_comparitor (const void* lhs, const void* rhs)
-{
+int pair_comparitor (const void* lhs, const void* rhs) {
     return (((pair *)lhs)->col - ((pair *)rhs)->col);
 }
 
 void init_A(char *filename_A, Data_Info info){
-    printf("begin to init A!\n");
-    A_ptr[0] = 0;
     int input_fd = open(filename_A, O_RDONLY);
-    ll size = info.nx * info.ny * info.nz;
+    ll size = info.tX * info.tY * info.tZ;
     around_point *data = (around_point *)malloc(size * sizeof(around_point));
-    if ((read(input_fd,data,sizeof(around_point)*size)) > 0){
+    A_ptr[0] = 0;
+    if ((read(input_fd,data,sizeof(around_point) * size)) > 0){
         int i, j, k;
         ll size_cube = info.ny * info.nz;
         ll size_all = info.tY * info.tZ;
@@ -181,24 +347,33 @@ void init_A(char *filename_A, Data_Info info){
                     int sub_num;
                     for (sub_num = 0; sub_num < 19; sub_num++){
                         double val = data[index_all].val[sub_num];
-                        int col = get_col(i + x_cor[sub_num], j + y_cor[sub_num], k + z_cor[sub_num], info);
+                        int g_index_v;
+                        int col = get_col(i + x_cor[sub_num], j + y_cor[sub_num], k + z_cor[sub_num], info, &g_index_v);
+
                         if (col >= 0){
                             data_tmp[pair_count].val = val;
                             data_tmp[pair_count].col = col;
 
+                            if (col >= size_cube * info.nx){
+                                g_map[col - size_cube * info.nx] = g_index_v;
+                            }
                             pair_count++;
                         }
                     }
                     // begin to order the pair
                     qsort(data_tmp, pair_count, sizeof(pair), pair_comparitor);
                     //begin to fill into A
+
                     for (sub_num = 0; sub_num < pair_count; sub_num++){
                         A_value[count] = data_tmp[sub_num].val;
                         A_col[count] = data_tmp[sub_num].col;
-                        //printf("value : %.6lf , col : %d \n", A_value[count], A_col[count]);
                         count++;
                     }
-
+                    if (info.myrank == 0){
+                        if (index_cube == 1231198){
+                            printf("\n");
+                        }
+                    }
                     A_ptr[index_cube + 1] = A_ptr[index_cube] + pair_count;
                 }
             }
@@ -206,13 +381,43 @@ void init_A(char *filename_A, Data_Info info){
     } else {
         printf("read A file error!\n");
     }
-
+    //free(data);
     close(input_fd);
+}
+
+void printf_x(Data_Info info){
+    int i, j, k;
+    //send side 0
+    int nx = info.nx;
+    int ny = info.ny;
+    int nz = info.nz;
+    int myrank = info.myrank;
+    int size_yz = ny * nz;
+    int start_x[8] = {0, 0, 0, 0, nx - 1, nx - 1, nx - 1, 0};
+    int end_x[8] =   {1, 1, 1, nx, nx, nx, nx, nx};
+    int start_y[8] = {0, 0, ny - 1, ny - 1, ny - 1, 0, 0, 0};
+    int end_y[8] =   {1, ny, ny, ny, ny, ny, 1, 1};
+    int side;
+    int size = info.nx * info.ny * info.nz;
+    int count = 0;
+    for (side = 0; side < 8; side++){
+        // 需要发送
+        printf("-------------side %d \n", side);
+        for (i = start_x[side]; i < end_x[side]; i++){
+            for (j = start_y[side]; j < end_y[side]; j++){
+                for (k = 0; k < nz; k++){
+                    double val = x[count + size];
+                    printf("side %d, i : %d, j : %d, k : %d, col : %d, val : %.10lf \n", side, i, j, k, count + size, val);
+                    count++;
+                }
+            }
+        }
+    }
 }
 
 void init_b(char *filename_b, Data_Info info){
     int input_fd = open(filename_b, O_RDONLY);
-    ll size = info.nx * info.ny * info.nz;
+    ll size = info.tX * info.tY * info.tZ;
     v_struct *data = (v_struct *)malloc(size * sizeof(v_struct));
     if ((read(input_fd,data,sizeof(around_point)*size)) > 0){
         int i, j, k;
@@ -237,15 +442,15 @@ void init_b(char *filename_b, Data_Info info){
 
 void init_x(char *filename_x, Data_Info info){
     int input_fd = open(filename_x, O_RDONLY);
-    ll size = info.nx * info.ny * info.nz;
+    ll size = info.tX * info.tY * info.tZ;
     v_struct *data = (v_struct *)malloc(size * sizeof(v_struct));
     if ((read(input_fd,data,sizeof(around_point)*size)) > 0){
         int i, j, k;
         ll size_cube = info.ny * info.nz;
         ll size_all = info.tY * info.tZ;
-        for (i = 0; i < info.nx; i++){
-            for (j = 0; j < info.ny; j++){
-                for (k = 0; k < info.nz; k++){
+        for (i = 0; i < info.nx; i++) {
+            for (j = 0; j < info.ny; j++) {
+                for (k = 0; k < info.nz; k++) {
                     ll index_cube = i * size_cube + j * info.nz + k; //row number of cube
                     ll index_all = (i + info.x_start) * size_all + (j + info.y_start) * info.tZ + k + info.z_start;
                     //fill into x_0 in x y z order
@@ -253,36 +458,7 @@ void init_x(char *filename_x, Data_Info info){
                 }
             }
         }
-        // begin to deal side
-        int count = info.nx * info.ny * info.nz;
-        //line 0
-        if (node_dsize[0] != -1){
 
-        }
-        if (node_dsize[1] != -1){
-
-        }
-        if (node_dsize[2] != -1){
-
-        }
-        if (node_dsize[3] != -1){
-
-        }
-        if (node_dsize[4] != -1){
-
-        }
-        if (node_dsize[5] != -1){
-
-        }
-        if (node_dsize[6] != -1){
-
-        }
-        if (node_dsize[7] != -1){
-
-        }
-        if (node_dsize[8] != -1){
-
-        }
     } else {
         printf("read A file error!\n");
     }
@@ -296,9 +472,6 @@ void A_m_vector(double * vector, int mrow, int mcol, double * result){
         int j_end = A_ptr[i + 1];
         double tmp = 0;
         for (j = j_start; j < j_end; j++){
-            if (A_col[j] >= mcol){
-                break;
-            }
             tmp += A_value[j] * vector[A_col[j]];
         }
         result[i] = tmp;
@@ -311,9 +484,6 @@ double get_value(int i, int j, double *matrix){
     int index;
     for (index = start_j; index < end_j; index++){
         if (A_col[index] == j){
-            if (i == 116484 && j == 130125){
-                printf("get value %.10lf \n", matrix[index]);
-            }
             return matrix[index];
         } else {
             if (A_col[index] > j){
@@ -332,32 +502,25 @@ void ILU_0(double *M, int rows){
         int start_pos = A_ptr[i];
         int end_pos = A_ptr[i + 1];
         for (tmp_pos = start_pos; tmp_pos < end_pos; tmp_pos++){
-            int k = A_col[tmp_pos];
-            if (k >= i - 1){
+            int j = A_col[tmp_pos];
+            if (j > i - 1){
                 break;
             }
-            double a_kk = get_value(k, k, M);
-            //printf("a_{%d, %d} : %.12lf \n", k, k, a_kk);
-            double a_ik = M[tmp_pos];
-            /*
-            if (i == 123323 && k == 130125){
-                printf("pos : %d \n", tmp_pos);
-                printf("a_{%d, %d} : %.12lf  A: %.12lf\n", i, k, M[tmp_pos], A_value[tmp_pos]);
-            }
-            */
-            a_ik = a_ik / a_kk;
-            M[tmp_pos] = a_ik;
+            double a_jj = get_value(j, j, M);
+            double a_ij = M[tmp_pos];
+            a_ij = a_ij / a_jj;
+            M[tmp_pos] = a_ij;
             int tmp_jpos;
             for(tmp_jpos = tmp_pos + 1; tmp_jpos < end_pos; tmp_jpos++){
-                int j = A_col[tmp_jpos];
-                if (j >= rows){
+                int k = A_col[tmp_jpos];
+                if (k >= rows){
                     //只使用方阵
                     break;
                 }
-                double a_ij = M[tmp_jpos];
+                double a_ik = M[tmp_jpos];
 
-                a_ij = a_ij - a_ik * get_value(k, j, M);
-                M[tmp_jpos] = a_ij;
+                a_ik = a_ik - a_ij * get_value(j, k, M);
+                M[tmp_jpos] = a_ik;
 
             }
         }
@@ -367,8 +530,7 @@ void ILU_0(double *M, int rows){
 void ilu_solver(double *R, double *M, double *R_hat, int rows){
     int i;
     // 第一轮
-    R_hat[0] = R[0];
-    for (i = 1; i < rows; i++){
+    for (i = 0; i < rows; i++){
         int start_j = A_ptr[i];
         int end_j = A_ptr[i + 1];
         int j;
@@ -400,6 +562,7 @@ void ilu_solver(double *R, double *M, double *R_hat, int rows){
     }
 }
 
+
 double vector_dot(double *vec1, double *vec2, int len){
     double result = 0;
     int i;
@@ -416,23 +579,33 @@ void update_x(double *p_i, double alpha, int len){
     }
 }
 
-int check_x(int rows){
+int check_x(int s, int rows, int myrank, Data_Info info){
     //需要进行 局部通讯以及 allreduce
     int i, j;
     double sum = 0;
-    for (i = 0; i < rows; i++){
+
+    for (i = s; i < rows; i++){
         int start_j = A_ptr[i];
         int end_j = A_ptr[i + 1];
         double b_i = b[i];
+        //double b_i = 0;
         double Ax_i = 0;
         for (j = start_j; j < end_j; j++){
             Ax_i += A_value[j] * x[A_col[j]];
         }
         sum += (b_i - Ax_i) * (b_i - Ax_i);
     }
-    sum = sqrt(sum);
-    printf("sum : %.10lf \n", sum);
-    if (sum < 0.000018789184606079){
+
+    double g_sum;
+    if(myrank == 0){
+        printf("sum : %.10lf \n", sum);
+    }
+    MPI_Allreduce(&sum, &g_sum, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+    g_sum = sqrt(g_sum);
+    if(myrank == 0){
+        printf("g_sum : %.10lf \n", g_sum);
+    }
+    if (g_sum < 0.000018789184606079){
         return 1;
     } else {
         return 0;
@@ -445,9 +618,6 @@ void check_R(int rows, double *R){
     for (i = 0; i < rows ; i++){
         tmp += R[i] * R[i];
     }
-    tmp = sqrt(tmp);
-
-    printf("R sum : %.10lf \n", tmp);
 }
 
 void check_P(int rows, double *P){
@@ -457,7 +627,7 @@ void check_P(int rows, double *P){
         tmp += P[i] * P[i];
     }
     tmp = sqrt(tmp);
-    printf("P sum : %.10lf \n", tmp);
+    //printf("P sum : %.10lf \n", tmp);
 }
 
 void update_R(double *R, double alpha, double *Ap_i, int len){
@@ -467,52 +637,148 @@ void update_R(double *R, double alpha, double *Ap_i, int len){
     }
 }
 
+void send_data(double *vector, int nx, int ny, int nz, int myrank){
+    int i, j, k;
+    //send side 0
+    int size_yz = ny * nz;
+    int start_x[8] = {0, 0, 0, 0, nx - 1, nx - 1, nx - 1, 0};
+    int end_x[8] =   {1, 1, 1, nx, nx, nx, nx, nx};
+    int start_y[8] = {0, 0, ny - 1, ny - 1, ny - 1, 0, 0, 0};
+    int end_y[8] =   {1, ny, ny, ny, ny, ny, 1, 1};
+    int side;
+
+    for (side = 0; side < 8; side++){
+        if (myrank != node_rank[side]){
+            // 需要发送
+            double *start_index = send_buffer + buffer_offset[side];
+            int count = 0;
+            for (i = start_x[side]; i < end_x[side]; i++){
+                for (j = start_y[side]; j < end_y[side]; j++){
+                    for (k = 0; k < nz; k++){
+                        int index = i * size_yz + j * nz + k;
+                        start_index[count] = vector[index];
+                        count++;
+                    }
+                }
+            }
+            // send data
+            MPI_Isend(start_index, node_dsize[side], MPI_DOUBLE, node_rank[side], target_side[side], MPI_COMM_WORLD, &send_request[side]);
+            //printf("send ------ side : %d, node_dsize : %d, target side : %d\n", side, node_dsize[side], target_side[side]);
+        }
+    }
+}
+
+void recv_data(double *vector, int myrank){
+    int side;
+    for (side = 0; side < 8; side++){
+        if (node_rank[side] != myrank){
+            double *receive_buffer = vector + buffer_offset[side];
+            MPI_Irecv(receive_buffer, node_dsize[side], MPI_DOUBLE, node_rank[side], side, MPI_COMM_WORLD, &recv_request[side]);
+            //printf("recv ------- side : %d, node_dsize : %d, side : %d \n", side, node_dsize[side], side);
+        }
+    }
+}
+
+void wait_send(int myrank){
+    int side;
+    MPI_Status status;
+    for (side = 0; side < 8; side++){
+        if (node_rank[side] != myrank){
+            MPI_Wait(&send_request[side], &status);
+        }
+    }
+}
+
+void wait_recv(int myrank){
+    int side;
+    MPI_Status status;
+    for (side = 0; side < 8; side++){
+        if (node_rank[side] != myrank){
+            MPI_Wait(&recv_request[side], &status);
+        }
+    }
+}
+
+void print_r_hat(double *v, int len){
+    int i;
+    for(i = 0; i < len; i++){
+        printf("rows : %d, R_hat : %.20lf \n", i, v[i]);
+    }
+}
+
+void check_sum_p(double *p, int len){
+    int i;
+    double tmp = 0;
+    for (i = 0; i < len; i++){
+        tmp += p[i] * p[i];
+    }
+    double g_sum;
+    MPI_Allreduce(&tmp, &g_sum, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+    printf("---------------------check sum : %.10lf \n", g_sum);
+}
+
 void gcr(Data_Info info, int k){
-    printf("begin to cal gcr\n");
+    //printf("begin to cal gcr\n");
     //init data
-    double *R = (double *) malloc(info.len_vb * sizeof(double));
+    double *R = (double *) malloc(info.len_x * sizeof(double));
+    if (R == NULL){
+        printf("R null \n");
+    }
     double *R_hat = (double *) malloc(info.len_x * sizeof(double)); // receive buffer is behind local values
+    if (R_hat == NULL){
+        printf("R_hat null \n");
+    }
     memset(R_hat, 0, info.len_x * sizeof(double));
     double *Ap = (double *) malloc(k * info.len_vb * sizeof(double));
+    if (Ap == NULL){
+        printf("Ap null \n");
+    }
     double *p = (double *) malloc(k * info.len_vb * sizeof(double));
+    if (p == NULL){
+        printf("p null \n");
+    }
     double *M = (double *) malloc(info.len_vb * 19 * sizeof(double));
+    if (M == NULL){
+        printf("M null \n");
+    }
     memcpy(M, A_value, info.len_vb * 19 * sizeof(double));
     //中间向量
     double *mid_result = (double *) malloc(info.len_vb * sizeof(double));
+    if (mid_result == NULL){
+        printf("mid_result null \n");
+    }
+    double * recv_start;
     //init R
+    //ssend and receive x
+    send_data(x, info.nx, info.ny, info.nz, info.myrank);
+    recv_start = x + info.len_vb;
+    recv_data(recv_start, info.myrank);
+    //waite recv done
+    wait_recv(info.myrank);
+    check_x(0, info.len_vb, info.myrank, info);
+
     A_m_vector(x, info.len_vb, info.len_x, R); //ok
+
     int i;
     for (i = 0; i < info.len_vb; i++){
         R[i] = b[i] - R[i];
     }
-    printf("init R done\n");
+    //check_R(info.len_vb, R);
     //init M
 
-    int tmp_p;
-    for (tmp_p = 0; tmp_p < info.len_vb; tmp_p++){
-        int s = A_ptr[tmp_p];
-        int e = A_ptr[tmp_p + 1];
-        int j;
-        //printf("row : %d \n", tmp_p);
-        for (j = s; j < e; j++){
-
-            if (A_col[j] == 130125 && tmp_p == 130162){
-                printf("A_%d%d : %.20lf\n", tmp_p, A_col[j], A_value[j]);
-                printf("M_%d%d : %.20lf\n", tmp_p, A_col[j], M[j]);
-                //printf("col: %d \n", tmp_p);
-            }
-
-            //printf("row: %d, col : %d \n", tmp_p, A_col[tmp_p]);
-        }
-    }
     ILU_0(M, info.len_vb);
-    printf("init M done\n");
+
     // get r hat
     ilu_solver(R, M, R_hat, info.len_vb);
-    printf("init R hat done\n");
+
     // init p_0
     memcpy(p, R_hat, info.len_vb * sizeof(double));
     // init Ap,  A * p_0 = A * R_hat , need communicate -------------------
+    send_data(R_hat, info.nx, info.ny, info.nz, info.myrank);
+    recv_start = R_hat + info.len_vb;
+    recv_data(recv_start, info.myrank);
+    wait_recv(info.myrank);
+
     A_m_vector(R_hat, info.len_vb, info.len_x, Ap);
     printf("init Ap done\n");
     int steps = 0; //迭代的次数
@@ -527,10 +793,16 @@ void gcr(Data_Info info, int k){
     while(1){
 
         //计算alpha 需要allreduce, R 需要局部通讯-------------------------------
-        double numerator = vector_dot(R, Ap_i, info.len_vb);
-        printf(" alpha numerator: %.30lf\n", numerator);
-        double denominator = vector_dot(Ap_i, Ap_i, info.len_vb);
+        double l_numerator = vector_dot(R, Ap_i, info.len_vb);
+        printf(" alpha numerator_l: %.30lf\n", l_numerator);
+        double l_denominator = vector_dot(Ap_i, Ap_i, info.len_vb);
+        printf(" alpha denominator local: %.10lf\n", l_denominator);
+        double denominator;
+        MPI_Allreduce(&l_denominator, &denominator, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+        double numerator;
+        MPI_Allreduce(&l_numerator, &numerator, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
         printf(" alpha denominator: %.10lf\n", denominator);
+        printf(" alpha numerator: %.10lf\n", numerator);
         Ap_idot[steps % k] = denominator;
         double alpha = numerator / denominator;
         printf("cal alpha done, alpha: %.10lf\n", alpha);
@@ -538,39 +810,42 @@ void gcr(Data_Info info, int k){
         update_x(p_i, alpha, info.len_vb);
         //printf("update x done\n");
         //开始check需要进行通讯---------------------
-        int judge = check_x(info.len_vb);
+        send_data(x, info.nx, info.ny, info.nz, info.myrank);
+        recv_start = x + info.len_vb;
+        recv_data(recv_start, info.myrank);
+        //waite recv done
+        wait_recv(info.myrank);
+        int judge = check_x(0, info.len_vb, info.myrank, info);
+
         //printf("check x done\n");
         if (judge == 1){
+            printf("step: %d, x get\n", steps);
             break;
         } else {
             printf("step: %d, not converge\n", steps);
         }
         //开始更新R
         update_R(R, alpha, Ap_i, info.len_vb);
-        //check_R(info.len_vb, R);
-        //printf("update R done\n");
         //跟新 R_hat 不需要通讯
         ilu_solver(R, M, R_hat, info.len_vb);
         //R_hat = R;
-        //printf("update R hat done\n");
         // 开始计算beta
         int sub_j;
         //计算 A * R_hat, 需要局部通讯到R_hat-----------------
+        send_data(R_hat, info.nx, info.ny, info.nz, info.myrank);
+        recv_start = R_hat + info.len_vb;
+        recv_data(recv_start, info.myrank);
+        wait_recv(info.myrank);
         A_m_vector(R_hat, info.len_vb, info.len_x, mid_result);
+
         for (sub_j = (steps / k) * k; sub_j <= steps; sub_j++){
             //计算分子的点积, 需要allreduce-------------------------
             double *Ap_j = Ap + (sub_j % k) * info.len_vb;
-            numerator = vector_dot(mid_result, Ap_j, info.len_vb);
+            l_numerator = vector_dot(mid_result, Ap_j, info.len_vb);
+            MPI_Allreduce(&l_numerator, &numerator, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
             denominator = Ap_idot[sub_j % k];
             beta[sub_j % k] = -numerator / denominator;
         }
-        /*
-        for (sub_j = 0; sub_j < k; sub_j++){
-            printf("beta[%d] : %0.19lf; ", sub_j, beta[sub_j]);
-        }
-        */
-        printf("\n");
-        //printf("get beta done\n");
         // 开始更新p_i
         steps++;
         p_i = p + (steps % k) * info.len_vb;
@@ -601,8 +876,11 @@ void gcr(Data_Info info, int k){
             Ap_i[sub_i] = tmp_result;
         }
         printf("update Ap_i done\n");
+        if (steps == 150){
+            printf("failed \n");
+            break;
+        }
     }
-
 }
 
 int main(int argc, char **argv) {
@@ -611,19 +889,21 @@ int main(int argc, char **argv) {
     int NX = atoi(argv[1]);
     int NY = atoi(argv[2]);
     int NZ = atoi(argv[3]);
-    int PX = atoi(argv[4]);
-    int PY = atoi(argv[5]);
-    int PZ = atoi(argv[6]);
+
     char *filename_A = argv[7];
     char *filename_x0 = argv[8];
     char *filename_b = argv[9];
      */
+    int nprocs;
+    MPI_Init(&argc, &argv);
+    MPI_Comm_size(MPI_COMM_WORLD, &nprocs);
     int NX = 360;
     int NY = 180;
     int NZ = 38;
-    int PX = 1;
-    int PY = 1;
-    int PZ = 1;
+    int PX = atoi(argv[1]);
+    int PY = atoi(argv[2]);
+    int PZ = atoi(argv[3]);
+    total_nodes = PX * PY * PZ;
     char *filename_A = "./data/case_1bin/data_A.bin";
     char *filename_x0 = "./data/case_1bin/data_x0.bin";
     char *filename_b = "./data/case_1bin/data_b.bin";
@@ -632,13 +912,20 @@ int main(int argc, char **argv) {
     }
     total_nodes = PX * PY * PZ;
     Data_Info info = init_info(NX, NY, NZ, PX, PY, PZ);
-    printf("init info done!\n");
+    //printf("init info done!\n");
     init_A(filename_A, info);
-    printf("init A done!\n");
+    //printf("init A done!\n");
     init_b(filename_b, info);
-    printf("init b done!\n");
+    //printf("init b done!\n");
     init_x(filename_x0, info);
-    printf("init x done!\n");
+    //printf("init x done!\n");
+    struct timeval t1, t2;
+    MPI_Barrier(MPI_COMM_WORLD), gettimeofday(&t1, NULL);
     gcr(info, 5);
+    MPI_Barrier(MPI_COMM_WORLD), gettimeofday(&t2, NULL);
+    if (info.myrank == 0) {
+        printf("Total time: %.6lf\n", TIME(t1, t2));
+    }
+    MPI_Finalize();
     return 0;
 }
